Revert "Allow login over plain http in restricted situations (#7835) (#8319)"
This reverts commit 46c3bd01f092caeba329f787790ce9f09af2fc42.
