+++ lib/utils/addr.go
@@ -23,7 +23,6 @@ import (
 	"strconv"
 	"strings"
 
-	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/trace"
 	log "github.com/sirupsen/logrus"
 )
@@ -77,7 +76,7 @@ func (a *NetAddr) IsLocal() bool {
 
 // IsLoopback returns true if this is a loopback address
 func (a *NetAddr) IsLoopback() bool {
-	return apiutils.IsLoopback(a.Addr)
+	return IsLoopback(a.Addr)
 }
 
 // IsHostUnspecified returns true if this address' host is unspecified.
@@ -270,6 +269,28 @@ func IsLocalhost(host string) bool {
 	}
 	ip := net.ParseIP(host)
 	return ip.IsLoopback() || ip.IsUnspecified()
+}
+
+// IsLoopback returns 'true' if a given hostname resolves to local
+// host's loopback interface
+func IsLoopback(host string) bool {
+	if strings.Contains(host, ":") {
+		var err error
+		host, _, err = net.SplitHostPort(host)
+		if err != nil {
+			return false
+		}
+	}
+	ips, err := net.LookupIP(host)
+	if err != nil {
+		return false
+	}
+	for _, ip := range ips {
+		if ip.IsLoopback() {
+			return true
+		}
+	}
+	return false
 }
 
 // GuessIP tries to guess an IP address this machine is reachable at on the
